Cвой вариант Словаря с использованием MVP.
