+++ app/src/main/java/by/romanovich/mydictinary/data/AppState.kt
@@ -1,0 +1,8 @@
+package by.romanovich.mydictinary.data
+
+sealed class AppState {
+    data class Success(val data: List<DataModel>?) : AppState()
+    data class Error(val error: Throwable) : AppState()
+    data class Loading(val progress: Int?) : AppState()
+}
+
