+++ app/src/main/java/by/romanovich/mydictinary/data/DataModel.kt
@@ -1,0 +1,11 @@
+package by.romanovich.mydictinary.data
+
+import com.google.gson.annotations.SerializedName
+
+
+class DataModel(
+    @field:SerializedName("text") val text: String?,
+    @field:SerializedName("meanings") val meanings: List<Meanings>?
+)
+
+
