+++ app/src/main/java/by/romanovich/mydictinary/data/Meanings.kt
@@ -1,0 +1,8 @@
+package by.romanovich.mydictinary.data
+
+import com.google.gson.annotations.SerializedName
+
+class Meanings(
+    @field:SerializedName("translation") val translation: Translation?,
+    @field:SerializedName("imageUrl") val imageUrl: String?
+)
