+++ app/src/main/java/by/romanovich/mydictinary/data/RepositoryImplementation.kt
@@ -1,0 +1,15 @@
+package by.romanovich.mydictinary.data
+
+import by.romanovich.mydictinary.data.DataModel
+import by.romanovich.mydictinary.domain.datasource.DataSource
+import by.romanovich.mydictinary.domain.repository.Repository
+import io.reactivex.Observable
+
+class RepositoryImplementation(private val dataSource: DataSource<List<DataModel>>) :
+    Repository<List<DataModel>> {
+    // Репозиторий возвращает данные, используя dataSource (локальный или
+// внешний)
+    override fun getData(word: String): Observable<List<DataModel>> {
+        return dataSource.getData(word)
+    }
+}
