+++ app/src/main/java/by/romanovich/mydictinary/data/Translation.kt
@@ -1,0 +1,5 @@
+package by.romanovich.mydictinary.data
+
+import com.google.gson.annotations.SerializedName
+
+class Translation(@field:SerializedName("text") val translation: String?)
