+++ app/src/main/java/by/romanovich/mydictinary/domain/datasource/ApiService.kt
@@ -1,0 +1,11 @@
+package by.romanovich.mydictinary.domain.datasource
+
+import by.romanovich.mydictinary.data.DataModel
+import io.reactivex.Observable
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ApiService {
+    @GET("words/search")
+    fun search(@Query("search") wordToSearch: String): Observable<List<DataModel>>
+}
