+++ app/src/main/java/by/romanovich/mydictinary/domain/datasource/DataSource.kt
@@ -1,0 +1,8 @@
+package by.romanovich.mydictinary.domain.datasource
+
+import io.reactivex.Observable
+
+// Источник данных для репозитория (Интернет, БД и т. п.)
+interface DataSource<T> {
+    fun getData(word: String): Observable<T>
+}
