+++ app/src/main/java/by/romanovich/mydictinary/domain/repository/Repository.kt
@@ -1,0 +1,9 @@
+package by.romanovich.mydictinary.domain.repository
+
+import io.reactivex.Observable
+
+// Репозиторий представляет собой слой получения и хранения данных, которые он
+// передаёт интерактору
+interface Repository<T> {
+    fun getData(word: String): Observable<T>
+}
