+++ app/src/main/java/by/romanovich/mydictinary/domain/rx/ISchedulerProvider.kt
@@ -1,0 +1,11 @@
+package by.romanovich.mydictinary.domain.rx
+
+import io.reactivex.Scheduler
+
+//In the sake of testing
+interface ISchedulerProvider {
+
+    fun ui(): Scheduler
+
+    fun io(): Scheduler
+}
