+++ app/src/main/java/by/romanovich/mydictinary/domain/rx/SchedulerProvider.kt
@@ -1,0 +1,12 @@
+package by.romanovich.mydictinary.domain.rx
+
+import io.reactivex.Scheduler
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
+
+class SchedulerProvider : ISchedulerProvider {
+
+    override fun ui(): Scheduler = AndroidSchedulers.mainThread()
+
+    override fun io(): Scheduler = Schedulers.io()
+}
