--- app/src/main/java/by/romanovich/mydictinary/model/data/DataModel.kt
@@ -1,11 +1,0 @@
-package by.romanovich.mydictinary.model.data
-
-import com.google.gson.annotations.SerializedName
-
-
-class DataModel(
-    @field:SerializedName("text") val text: String?,
-    @field:SerializedName("meanings") val meanings: List<Meanings>?
-)
-
-
