--- app/src/main/java/by/romanovich/mydictinary/model/data/Translation.kt
@@ -1,5 +1,0 @@
-package by.romanovich.mydictinary.model.data
-
-import com.google.gson.annotations.SerializedName
-
-class Translation(@field:SerializedName("text") val translation: String?)
