--- app/src/main/java/by/romanovich/mydictinary/model/datasource/ApiService.kt
@@ -1,11 +1,0 @@
-package by.romanovich.mydictinary.model.datasource
-
-import by.romanovich.mydictinary.model.data.DataModel
-import io.reactivex.Observable
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ApiService {
-    @GET("words/search")
-    fun search(@Query("search") wordToSearch: String): Observable<List<DataModel>>
-}
