--- app/src/main/java/by/romanovich/mydictinary/model/datasource/DataSource.kt
@@ -1,8 +1,0 @@
-package by.romanovich.mydictinary.model.datasource
-
-import io.reactivex.Observable
-
-// Источник данных для репозитория (Интернет, БД и т. п.)
-interface DataSource<T> {
-    fun getData(word: String): Observable<T>
-}
