--- app/src/main/java/by/romanovich/mydictinary/model/datasource/DataSourceLocal.kt
@@ -1,13 +1,0 @@
-package by.romanovich.mydictinary.model.datasource
-
-import by.romanovich.mydictinary.model.data.DataModel
-import io.reactivex.Observable
-
-// Для локальных данных используется Room
-class DataSourceLocal(private val remoteProvider: RoomDataBaseImplementation =
-                          RoomDataBaseImplementation()
-) :
-    DataSource<List<DataModel>> {
-    override fun getData(word: String): Observable<List<DataModel>> =
-        remoteProvider.getData(word)
-}
