--- app/src/main/java/by/romanovich/mydictinary/model/datasource/DataSourceRemote.kt
@@ -1,13 +1,0 @@
-package by.romanovich.mydictinary.model.datasource
-
-import by.romanovich.mydictinary.model.data.DataModel
-import io.reactivex.Observable
-
-// Для получения внешних данных мы будем использовать Retrofit
-class DataSourceRemote(private val remoteProvider: RetrofitImplementation =
-    RetrofitImplementation()
-) : DataSource<List<DataModel>> {
-    override fun getData(word: String): Observable<List<DataModel>> =
-        remoteProvider.getData(word)
-}
-
