--- app/src/main/java/by/romanovich/mydictinary/model/datasource/RetrofitImplementation.kt
@@ -1,38 +1,0 @@
-package by.romanovich.mydictinary.model.datasource
-
-import by.romanovich.mydictinary.model.data.DataModel
-import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import io.reactivex.Observable
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class RetrofitImplementation : DataSource<List<DataModel>> {
-    override fun getData(word: String): Observable<List<DataModel>> {
-        return getService(BaseInterceptor.interceptor).search(word)
-    }
-    private fun getService(interceptor: Interceptor): ApiService {
-        return createRetrofit(interceptor).create(ApiService::class.java)
-    }
-    private fun createRetrofit(interceptor: Interceptor): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(BASE_URL_LOCATIONS)
-            .addConverterFactory(GsonConverterFactory.create())
-            .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-            .client(createOkHttpClient(interceptor))
-            .build()
-    }
-    private fun createOkHttpClient(interceptor: Interceptor): OkHttpClient {
-        val httpClient = OkHttpClient.Builder()
-        httpClient.addInterceptor(interceptor)
-        httpClient.addInterceptor(
-            HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-        return httpClient.build()
-    }
-    companion object {
-        private const val BASE_URL_LOCATIONS =
-            "https://dictionary.skyeng.ru/api/public/v1/"
-    }
-}
