--- app/src/main/java/by/romanovich/mydictinary/model/repository/Repository.kt
@@ -1,9 +1,0 @@
-package by.romanovich.mydictinary.model.repository
-
-import io.reactivex.Observable
-
-// Репозиторий представляет собой слой получения и хранения данных, которые он
-// передаёт интерактору
-interface Repository<T> {
-    fun getData(word: String): Observable<T>
-}
