--- app/src/main/java/by/romanovich/mydictinary/presenter/Interactor.kt
@@ -1,10 +1,0 @@
-package by.romanovich.mydictinary.presenter
-
-import io.reactivex.Observable
-
-// Ещё выше стоит интерактор. Здесь уже чистая бизнес-логика
-interface Interactor<T> {
-    // Use Сase: получение данных для вывода на экран
-// Используем RxJava
-    fun getData(word: String, fromRemoteSource: Boolean): Observable<T>
-}
