--- app/src/main/java/by/romanovich/mydictinary/rx/ISchedulerProvider.kt
@@ -1,11 +1,0 @@
-package by.romanovich.mydictinary.rx
-
-import io.reactivex.Scheduler
-
-//In the sake of testing
-interface ISchedulerProvider {
-
-    fun ui(): Scheduler
-
-    fun io(): Scheduler
-}
