--- app/src/main/java/by/romanovich/mydictinary/rx/SchedulerProvider.kt
@@ -1,13 +1,0 @@
-package by.romanovich.mydictinary.rx
-
-import by.romanovich.mydictinary.rx.ISchedulerProvider
-import io.reactivex.Scheduler
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
-
-class SchedulerProvider : ISchedulerProvider {
-
-    override fun ui(): Scheduler = AndroidSchedulers.mainThread()
-
-    override fun io(): Scheduler = Schedulers.io()
-}
