+++ app/src/main/java/by/romanovich/mydictinary/ui/BaseFragment.kt
@@ -1,0 +1,32 @@
+package by.romanovich.mydictinary.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
+
+abstract class BaseFragment<VB : ViewBinding>(private val bindingInflater: (inflater: LayoutInflater) -> VB) :
+    Fragment() {
+
+    private var _binding: VB? = null
+    protected val binding: VB
+        get() = _binding as VB
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = bindingInflater.invoke(layoutInflater)
+        if (_binding == null)
+            throw IllegalAccessException("binding cannot be null")
+        return binding.root
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
+}
