+++ app/src/main/java/by/romanovich/mydictinary/ui/main/MainActivity.kt
@@ -1,0 +1,106 @@
+package by.romanovich.mydictinary.ui.main
+
+import android.os.Bundle
+import android.view.View.GONE
+import android.view.View.VISIBLE
+import android.widget.Toast
+import androidx.recyclerview.widget.LinearLayoutManager
+import by.romanovich.mydictinary.data.DataModel
+import by.romanovich.mydictinary.data.AppState
+import by.romanovich.mydictinary.ui.base.BaseActivity
+import by.romanovich.mydictinary.ui.main.adapter.MainAdapter
+import by.romanovich.mydictinary.R
+import by.romanovich.mydictinary.databinding.ActivityMainBinding
+import by.romanovich.mydictinary.ui.translator.TranslationFragment
+import by.romanovich.mydictinary.ui.translator.TranslatorContract
+import by.romanovich.mydictinary.ui.translator.TranslatorPresenterImpl
+
+class MainActivity : BaseActivity<AppState>() {
+    private lateinit var binding: ActivityMainBinding
+    private var adapter: MainAdapter? = null
+    private val onListItemClickListener: MainAdapter.OnListItemClickListener =
+        object : MainAdapter.OnListItemClickListener {
+            override fun onItemClick(data: DataModel) {
+                Toast.makeText(this@MainActivity, data.text,
+                    Toast.LENGTH_SHORT).show()
+            }
+        }
+    override fun createPresenter(): TranslatorContract.Presenter<AppState, TranslatorContract.View> {
+        return TranslatorPresenterImpl()
+    }
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.searchFab.setOnClickListener {
+            val translationFragment = TranslationFragment.newInstance()
+            translationFragment.setOnSearchClickListener(object :
+
+                TranslationFragment.OnSearchClickListener {
+                override fun onClick(searchWord: String) {
+                    presenter.getData(searchWord, true)
+                }
+            })
+            translationFragment.show(supportFragmentManager,
+                BOTTOM_SHEET_FRAGMENT_DIALOG_TAG
+            )
+        }
+    }
+    override fun renderData(appState: AppState) {
+        when (appState) {
+            is AppState.Success -> {
+                val dataModel = appState.data
+                if (dataModel == null || dataModel.isEmpty()) {
+                    showErrorScreen(getString(R.string.empty_server_response_on_success))
+                } else {
+                    showViewSuccess()
+                    if (adapter == null) {
+                        binding.mainActivityRecyclerview.layoutManager = LinearLayoutManager(applicationContext)
+                        binding.mainActivityRecyclerview.adapter = MainAdapter(onListItemClickListener, dataModel)
+                    } else {
+                        adapter!!.setData(dataModel)
+                    }
+                }
+            }
+            is AppState.Loading -> {
+                showViewLoading()
+                if (appState.progress != null) {
+                    binding.progressBarHorizontal.visibility = VISIBLE
+                    binding.progressBarRound.visibility = GONE
+                    binding.progressBarHorizontal.progress = appState.progress
+                } else {
+                    binding.progressBarHorizontal.visibility = GONE
+                    binding.progressBarRound.visibility = VISIBLE
+                }
+            }
+            is AppState.Error -> {
+                showErrorScreen(appState.error.message)
+            }
+        }
+    }
+    private fun showErrorScreen(error: String?) {
+        showViewError()
+        binding.errorTextview.text = error ?: getString(R.string.undefined_error)
+        binding.reloadButton.setOnClickListener {
+            presenter.getData("hi", true)
+        }
+    }
+    private fun showViewSuccess() {
+        binding.successLinearLayout.visibility = VISIBLE
+        binding.loadingFrameLayout.visibility = GONE
+        binding.errorLinearLayout.visibility = GONE
+    }
+    private fun showViewLoading() {
+        binding.successLinearLayout.visibility = GONE
+        binding.loadingFrameLayout.visibility = VISIBLE
+        binding.errorLinearLayout.visibility = GONE
+    }
+    private fun showViewError() {
+        binding.successLinearLayout.visibility = GONE
+        binding.loadingFrameLayout.visibility = GONE
+        binding.errorLinearLayout.visibility = VISIBLE
+    }
+    companion object {
+        private const val BOTTOM_SHEET_FRAGMENT_DIALOG_TAG = "74a54328-5d62-46bf-ab6b-cbf5fgt0-092395"
+    }
+}
