+++ app/src/main/java/by/romanovich/mydictinary/ui/main/adapter/MainAdapter.kt
@@ -1,0 +1,49 @@
+package by.romanovich.mydictinary.ui.main.adapter
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import by.romanovich.mydictinary.R
+import by.romanovich.mydictinary.data.DataModel
+
+
+class MainAdapter(
+    private var onListItemClickListener: OnListItemClickListener,
+    private var data: List<DataModel>
+) : RecyclerView.Adapter<MainAdapter.RecyclerItemViewHolder>() {
+    fun setData(data: List<DataModel>) {
+        this.data = data
+        notifyDataSetChanged()
+    }
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int):
+            RecyclerItemViewHolder {
+        return RecyclerItemViewHolder(
+            LayoutInflater.from(parent.context)
+                .inflate(R.layout.activity_main_translation_item, parent, false) as View
+        )
+    }
+    override fun onBindViewHolder(holder: RecyclerItemViewHolder, position: Int) {
+        holder.bind(data.get(position))
+    }
+    override fun getItemCount(): Int {
+        return data.size
+    }
+    inner class RecyclerItemViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+        fun bind(data: DataModel) {
+            if (layoutPosition != RecyclerView.NO_POSITION) {
+                itemView.findViewById<TextView>(R.id.header_textview_recycler_item).text = data.text
+                itemView.findViewById<TextView>(R.id.description_textview_recycler_item).text =
+                    data.meanings?.get(0)?.translation?.translation
+                itemView.setOnClickListener { openInNewWindow(data) }
+            }
+        }
+    }
+    private fun openInNewWindow(listItemData: DataModel) {
+        onListItemClickListener.onItemClick(listItemData)
+    }
+    interface OnListItemClickListener {
+        fun onItemClick(data: DataModel)
+    }
+}
