--- app/src/main/java/by/romanovich/mydictinary/view/main/MainInteractor.kt
@@ -1,25 +1,0 @@
-package by.romanovich.mydictinary.view.main
-
-import by.romanovich.mydictinary.model.data.DataModel
-import by.romanovich.mydictinary.model.data.AppState
-import by.romanovich.mydictinary.model.repository.Repository
-import by.romanovich.mydictinary.presenter.Interactor
-import io.reactivex.Observable
-
-class MainInteractor(
-// Снабжаем интерактор репозиторием для получения локальных или внешних
-// данных
-    private val remoteRepository: Repository<List<DataModel>>,
-    private val localRepository: Repository<List<DataModel>>
-) : Interactor<AppState> {
-    // Интерактор лишь запрашивает у репозитория данные, детали имплементации
-// интерактору неизвестны
-    override fun getData(word: String, fromRemoteSource: Boolean):
-            Observable<AppState> {
-        return if (fromRemoteSource) {
-            remoteRepository.getData(word).map { AppState.Success(it) }
-        } else {
-            localRepository.getData(word).map { AppState.Success(it) }
-        }
-    }
-}
