+++ app/src/main/res/layout/activity_main_translation_item.xml
