+++ app/src/main/res/layout/fragment_translation.xml
